feat: move wrappers from comonents to screens/wrappers
