+++ src/components/index.ts
@@ -1,4 +1,5 @@
 import Input from "./Input";
 import Rings from "./Rings";
+import { CustomModal } from "./modal";
 
-export { Input, Rings };
+export { Input, Rings, CustomModal as Modal };
